Update printLabel function

+++ gui/graphicalInterface/Text.java
@@ -1,9 +1,9 @@
 public class Text {
 	public String label;
-	public float x;
-	public float y;
+	public int x;
+	public int y;
 	
-	Text(String label, float x, float y){
+	Text(String label, int x, int y){
 		this.label = label;
 		this.x = x;
 		this.y = y;
